Have something working for CP. not sure if correct, but it's working.

+++ dpgen/src/DatapathComponent.cpp
@@ -67,7 +67,7 @@ void DatapathComponent::setDataWidth(int dW)
 	dataWidth = dW;
 }
 
-int DatapathComponent::getDelay()
+double DatapathComponent::getDelay()
 {
 	return delay;
 }
